* uvh5 header_t->uvw is double*

+++ include/uvh5.h
@@ -52,7 +52,7 @@ typedef struct
 	int Nfreqs;									 /* The total number of frequency channels in the data across all spectral windows. */
 	int Npols;									 /* The number of polarization products in the data. */
 	int Ntimes;									 /* The number of time samples present in the data. */
-	float *uvw_array;						 /* An array of the uvw-coordinates corresponding to each observation in the data. This is a
+	double *uvw_array;					 /* An array of the uvw-coordinates corresponding to each observation in the data. This is a
 																	two-dimensional array of size [Nbls, 3] (reiteratively written Ntimes, so that stored it
 																	has size [Nblts, 3]). Units are in meters. */
 	float *time_array;					 /* An array of the Julian Date corresponding to the temporal midpoint of the corresponding
